edited the close order endpoint

+++ Models/Order.cs
@@ -14,7 +14,7 @@ namespace HHPW.Models
         public int UserId { get; set; }
         public DateTime OrderTime { get; set; }
         public DateTime? CloseTime { get; set; }
-        public int RevTotal { get; set; }
+        public decimal RevTotal { get; set; }
         public int Tip {  get; set; }
         public string? PaymentType { get; set; }
         public List<OrderItem> Items { get; set; } = new List<OrderItem>();
